增加新版本检测，并提示升级
