+++ app.js
@@ -3,6 +3,20 @@ App({
   navigateToLogin: false,
   onLaunch: function() {
     const that = this;
+    // 检测新版本
+    const updateManager = wx.getUpdateManager()
+    updateManager.onUpdateReady(function () {
+      wx.showModal({
+        title: '更新提示',
+        content: '新版本已经准备好，是否重启应用？',
+        success(res) {
+          if (res.confirm) {
+            // 新的版本已经下载好，调用 applyUpdate 应用新版本并重启
+            updateManager.applyUpdate()
+          }
+        }
+      })
+    })
     /**
      * 初次加载判断网络情况
      * 无网络状态下根据实际情况进行调整
