+++ config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  version: "6.0.2",
+  version: "6.0.3",
   note: '接口封装',
   subDomain: "tz", // 如果你的域名是： https://api.it120.cc/abcd 那么这里只要填写 abcd
   appid: "wxa46b09d413fbcaff", // 您的小程序的appid，购物单功能需要使用
